requirements.txt
# Flask Core
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.23

# Image Processing
Pillow>=9.0.0

# Environment
python-dotenv==1.0.0

# Development (optional)
# pytest==7.4.3
# pytest-flask==1.3.0

# Production (optional)
# gunicorn==21.2.0